Refactor dockerfile
This commit will refactor the Dockerfile. Remove some redundant steps
and simplify the logic. To simplify the logic it will copy everything from the repo into the
image. This may cause some unnecessary files to be copied, on the other
hand it will reduce the need of maintenance if files are added.