Add docker support (#3)
* Added dockerfile

This dockerfile will successfully build a docker image for fits with the
command `docker build . -t fits`

* Reduce image size with docker scratch.

This commits reduces the image size by omitting libraries needed to
build to go binary. It uses docker scratch that creates an image
only with files needed to run fits.

* Added support for docker environment.

This commit adds an environment variable for docker to fits. This allows
fits to set a internal docker folder as watcher folders instead
of (unreachable) folders outside of the container.

* Update documentation how to run docker image

* Refactor fits environment to variable

* Add docker build to makefile

This commit adds the option to run docker build command from makefile by
running `make docker`

* Reduce docker image size

This commit adds ldflags -s and -w to the Dockerfile. This will omit debug information and
DWARF table from the output binary, reducing the binary size.

* Refactor dockerfile

This commit will refactor the Dockerfile. Remove some redundant steps
and simplify the logic. To simplify the logic it will copy everything from the repo into the
image. This may cause some unnecessary files to be copied, on the other
hand it will reduce the need of maintenance if files are added.

* Modify docker working directory

This commit will change the working directory for building the go
binary. This is in any case the golang-alpine image files will conflict with
the files in the repo in the future.

* Refactor to clearify build directory

* Update README.md

This commit clarify that we are building and running docker images.