Add ftp send support
This commit will add support to send files via ftp.
It also applies some refactor to separate logic for
senders and watchers. It introduces a common.Path
interface that contains information needed send files.
By using Gateway with a generic `Send` in the send-gateway.go
we can now send files in the file system or via ftp.
The vision is that the common.Path can be used to reference many
different transfer types.