Change to polling from fsnotify
This commit will use polling to check
for file changes every 10 seconds instead
of relying on fsnotify. This is because
fsnotify makes it hard to see when a file
is fully created, e.g. when a file is
streamed and written to multiple times.
Also we probably need polling anyway if
we want to check for changes from anything
other than file system.