Refactor read, send and remove logic
This commit will refactor the logic of
read, send and remove into gateways.
This will prepare for easier
implementation of transfer from other
types than file system, e.g. ftp & ssh
in the future.

Every time an operation is made it will
redirect it via the gateway. The gateway
will then check the scheme to determine
what implementation to use.