Refactor replace gateway with interface
This commit will replace the gateway with
an fileservice interface. It started to be
pretty tedious to check the scheme all
the time so an interface felt like a better
solution.

Now we only check it once to find
the correct implementation. The fileservice
will then implement calls for List, Read,
Send and Delete.