Add fits helm chart
This commit will add a helm chart for fits
that makes it easy to deploy to kubernetes.

If you're on a single node cluster you can
use file system paths. The chart will try
to set up a hostPath mount on the specified
path in your cluster. Fits will try to check
if you use any local paths (that starts with "/")