Bug helm path for smb would try to mount
This commit will fix that if you enter a smb
path (that starts with two "//") it will not
try to mount a hostPath for local file system.

Some hacky checks to accomplish this in helm
but I still value the ease of use of the chart.
We dont need to introduce a variable to manually
set type of path url.